deleted logs and fixed errors

+++ src/js/sweet.js
@@ -1,7 +1,7 @@
 //Store buttons
 const sweetAddDelay = () => {
     const addToCart = document.getElementsByClassName("addToCart");
-    
+
     for (let i = 0; i < addToCart.length; i++) {
         addToCart[i].addEventListener("click", function (event) {
             event.preventDefault();
@@ -9,22 +9,23 @@ const sweetAddDelay = () => {
                 title: "Item added!",
                 icon: "success",
                 timer: 800,
-                showConfirmButton: false
+                showConfirmButton: false,
             });
         });
     }
-}
+};
 setTimeout(sweetAddDelay, 1000);
 
 //Checkout button
 const checkoutButton = document.getElementById("checkoutButton");
-
-checkoutButton.addEventListener("click", function (event) {
-    event.preventDefault();
-    localStorage.clear();
-    Swal.fire({
-        title: "Checkout complete!",
-        icon: "success",
-        confirmButtonText: "Done",
+if (checkoutButton) {
+    checkoutButton.addEventListener("click", function (event) {
+        event.preventDefault();
+        localStorage.clear();
+        Swal.fire({
+            title: "Checkout complete!",
+            icon: "success",
+            confirmButtonText: "Done",
+        });
     });
-});
+}
